+++ lightningd/Makefile
@@ -73,7 +73,6 @@ LIGHTNINGD_SRC :=				\
 	lightningd/peer_control.c		\
 	lightningd/peer_htlcs.c			\
 	lightningd/subd.c			\
-	lightningd/txfilter.c			\
 	lightningd/watch.c
 
 # Source files without corresponding headers
