+++ wallet/Makefile
@@ -7,6 +7,7 @@ wallet-wrongdir:
 WALLET_LIB_SRC :=		\
 	wallet/db.c		\
 	wallet/invoices.c	\
+	wallet/txfilter.c	\
 	wallet/wallet.c		\
 	wallet/walletrpc.c
